move waymo lane property to constant
