+++ metadrive/component/map/waymo_map.py
@@ -6,7 +6,7 @@ from metadrive.component.waymo_block.waymo_block import WaymoBlock
 from metadrive.engine.asset_loader import AssetLoader
 from metadrive.type import MetaDriveType
 from metadrive.utils.waymo_utils.utils import convert_polyline_to_metadrive, read_waymo_data
-from metadrive.utils.waymo_utils.waymo_type import WaymoLaneProperty
+from metadrive.constants import WaymoLaneProperty
 
 
 class WaymoMap(BaseMap):
