+++ metadrive/constants.py
@@ -311,3 +311,13 @@ REPLAY_DONE = "replay_done"
 
 class NuPlanLaneProperty:
     pass
+
+
+class WaymoLaneProperty:
+    POLYLINE = "polyline"
+    LEFT_BOUNDARIES = "left_boundaries"
+    RIGHT_BOUNDARIES = "right_boundaries"
+    LEFT_NEIGHBORS = "left_neighbor"
+    RIGHT_NEIGHBORS = "right_neighbor"
+    ENTRY = "entry_lanes"
+    EXIT = "exit_lanes"
